+++ packages/manager/src/dev-tools/FeatureFlagTool.tsx
@@ -32,6 +32,7 @@ const options: { flag: keyof Flags; label: string }[] = [
   { flag: 'dbaasV2', label: 'Databases V2 Beta' },
   { flag: 'databaseResize', label: 'Database Resize' },
   { flag: 'apicliDxToolsAdditions', label: 'APICLI DX Tools Additions' },
+  { flag: 'apicliButtonCopy', label: 'APICLI Button Copy' },
 ];
 
 const renderFlagItems = (