+++ packages/manager/src/features/Linodes/LinodesCreate/CodeBlock/CodeBlock.tsx
@@ -1,6 +1,9 @@
+import { useLDClient } from 'launchdarkly-react-client-sdk';
 import React from 'react';
 
+import { useFlags } from 'src/hooks/useFlags';
 import { sendApiAwarenessClickEvent } from 'src/utilities/analytics/customEventAnalytics';
+
 import {
   StyledCommandDiv,
   StyledCopyTooltip,
@@ -11,13 +14,24 @@ export interface CodeBlockProps {
   command: string;
   commandType: string;
   language: 'bash';
+  ldTrackingKey?: string;
 }
 
 export const CodeBlock = (props: CodeBlockProps) => {
-  const { command, commandType, language } = props;
+  const flags = useFlags();
+  const ldClient = useLDClient();
+
+  const { command, commandType, language, ldTrackingKey } = props;
+
+  const apicliButtonCopy = flags?.testdxtoolabexperiment;
 
   const handleCopyIconClick = () => {
     sendApiAwarenessClickEvent('Copy Icon', commandType);
+    if (ldTrackingKey) {
+      ldClient?.track(ldTrackingKey, {
+        variation: apicliButtonCopy,
+      });
+    }
   };
 
   return (